core-package: remove unwanted optional package
The package 'xserver-xorg-input-libinput' was included as the touchpad
didn't function well in X server window manager on Debian. 'synclient'
was needed to configure the touchpad to work well. It was found in that
package.

It's not very useful as it's not guaranteed to happen all the time.
Also, the package is not guaranteed to have 'synclient' (the
'xserver-xorg-input-synaptics' once had the package instead of this.